Update from 1.8.9 conventions to 1.10.2
Changed the naming of references to vanilla blocks to the new constants.

Changed the import statements of several things, to properly point to
their new packages.

We now call a getter instead of referencing an object directly, in the
SpawnBlockEvent.

Changed wording on configuration entry.

Added a new reference and changed a call in the replacing statement
inside our WorldGenMineable, as the method from 1.8.9 no longer works.

Version number changed.

** REGISTER BLOCK CALL IS DEPRECATED **
I tried to update it right now, but that crashed the smelting recipe. I
don't have the time or energy to figure out how to work the new
registering method, so while it's just deprecated (and thus still
working) I won't bother. I'll try to update it when I get the chance.