Server Crash Fix Attempt #2
It seems the actual problem was actually that the method being marked as
Client Only wasn't enough, as it was still being called on the server
side, but then the server didn't know what it was.

Checking if the world is remote (aka is the client) before calling that
should fix things up.