Fixing Missing Class Crashes
Mekanism changed the location of their config class, which was causing a crash. Also added (tentative?) support for IC2 Classic, which while having the same mod ID as IC2 Original/Experimental, it has a completely different layout, so support had to be added in seperately. Should coexist nicely with old IC2.

Temporarily removed support for Advanced Rocketry, as I couldn't get it to load in the dev enviromnent, so I can't verify that it works.

Should have fixed future crashes if a class is removed... hopefuly....